refactor(internal/sidekick/gcloud): split generate.go by topic
The file had grown long and was mixing several jobs. Move the CLI
model types and the code that builds the model into model.go. Move
the Go client package lookup into client.go. Move the resource-path
helpers into path.go. What is left in generate.go is the top-level
flow: Generate, renderMain, writeMain, renderReadme.

Also rewrite constructCLIModel. The old version did three things in
one loop and used a flag to remember whether any command needed a
client import. The new version runs in three steps. First build the
commands, then bucket them into subgroups, then decide which imports
to add by checking the commands. The flag is gone. A small new helper
in path.go figures out the subgroup name for a method, so the main
loop does not have to.

Add doc comments to the fields of Group, Subgroup, Command, and
commandWithSubgroup.
